+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [0.0.34] - 2026-01-13
+
+### Added in 0.0.34
+
+- _sz_database.py support for AURORAPOSTGRESQL and new connection string parameters
+
+### Fixed in 0.0.34
+
+- Numerous small issues in _sz_database.py
+
+### Changes in 0.0.34
+
+- Method names in _sz_database.py to be pythonic
+- sz_explorer and sz_snapshot for changes to_sz_database.py 
+
 ## [0.0.32] - 2025-11-12
 
 ### Fixed in 0.0.32